Add to node origins while consuming pgn
Previously, we computed all origins for nodes in the graph after the
graph has been built from a pgn and saturated. However, this does not
make it possible to distinguish lines that can be found directly in the
source pgn from lines that can only be found in the built graph.

Therefore, we add to Node.origins already while building the graph from
the pgn. We also add meta data to the node that allows us to determine which
origins were found directly in a pgn and which origins were found in the
built graph. This will be useful for a situation where the user wants to
find a particular position in the original pgn.

Also added a function Node.print_origins.