change get_node method to raise ValueError
Previously, Graph.get_node returned None when there is no node
corresponding to the input fen. However, this turned out to be
impractical because the programmer cannot rely on get_node returning a
node. Therefore we changed it so that it always returns a Node or else
raises an error.

We also added a method node_exists for checking whether a node exists
for a given fen.