Add 'max_depth' and 'move_selection' parameters
These parameters affect practice mode and can be changed during practice
mode.

Each practice run ends when a depth of max_depth is reached. More
specifically, this means when the max_depth'th move has been played by
the user.

Previously, 'random-leaf' was the only available move_selection move by
which the opponent in practice mode selects her moves. Now, there is a
second move 'uniform' and 'uniform' was made the default.

The 'random-leaf' mode may be changed in the future so that it actually
ends at each leaf with the same probability. (i.e. selects a leaf at
random, then selects a line leading to that leaf at random).