Graph is now initialized with list of pgns (Games)
This will allow us to split our data across multiple pgn files in the
future in order to make it easier to keep track of changes.

I also changed statements of the form
pgn = open(filename))
to
with open(filename) as pgn:
in order to make sure that files are closed once they are no longer
needed.