Split data files into smaller files
Previously it was difficult to keep track of changes, because the whole
data file was rewritten whenever some data was added to it. Now it will
be easier to keep track of what was changed, because most files will
typically remain unchanged. In the future we may also add more files.
New data may be added in the form of "delta" files instead of updating
existing files.