Change random-leaf mode to pick random leaf
Completely change the logic for random leaf mode. Previously, it picked
each move randomly and the probability for picking a move was
proportional to the number of leaves in the subtree after that move.
However, the problem with this old method is that some leaves can be
reached through several different moves, and these leaves are reached
disproportionately often.

The new method picks a random leaf at the beginning, as well as a random
origin leading to that leaf, and the opponent then plays the moves
according to that origin string. Random-leaf mode is now also the
default.