Implement probability analysis of positions
Use lichess 'explore openings' api to calculate the probabilities of any position to come up which
is either inside the tree or one ply outside of the tree. The script analyze_coverage.py then
creates a report where all positions are ordered by their probability. This report can help
decide which positions to explore next in order to close gaps in the data or it can be used as
a guide for which positions to study