More catalan data including a line with loop
Line with loop is currently excluded from being used to build the
graph because the program is not designed to handle a graph with a
loop. This may be a direction for a future improvement.
However, all but the last two moves of
black_catalan_line_with_loop.pgn
are also in
black_vs_catalan.pgn
Note also: Since the evaluation is slightly negative (Q = -0.008) in the
position at the end of black_catalan_line_with_loop.pgn, repeating moves
and accepting the resulting draw would be advantageous for black.